src/log.c: fix segfault when using `errlog=`
The log.c:log_info() function caused a segfault by reusing a va_list on
a second vfprintf() call whenever the errlog parameter was used.

Fix consisted in using va_copy() to make a copy of the original va_list
to be used by the errorlog's vfprintf() call, and having a respective
va_end().

Closes: #5