src/sizeprobe.c: fix destination size unit
As reported by Tim Phipps [1] on the Debian Bug Tracking System, a
Debian-specific patch was applied to the Debian version of dcfldd, which
caused the get_dev_size function to use BLKGETSIZE64 instead of
BLKGETSIZE, without accounting for the difference between the units used
by each one (sectors vs bytes).

As that same patch was applied to the github version of dcfldd by commit
0f4bb57, the same behaviour was present on the github version. So this
fix returns the result from BLKGETSIZE64 without multiplying by 512.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=886647#10

Closes: #6