src/util.c: fix non-Linux build error
On Linux, skip_via_lseek() is a wrapper function for lseek() that works
around a specific kernel bug. On other platforms, it is just a macro
that calls lseek().

skip_via_lseek() uses the error() function available on the error.h
header. The functions declared on that header are GNU extensions and are
not meant to be portable. Thus, the global include breaks the build on
non-GNU based systems.

As the only usage of that function on the file is inside the above
mentioned function, this fix moves the #include inside the #ifdef
responsible for the function definition.

Tested under Debian GNU/Linux 10 (buster) and macOS 10.15.2 (Catalina).