Release `make distcheck` tarball on tag push
Uses GitHub Actions Workflow to build a release tarball which doesn't
need autotools to configure and make.

To release using this feature, GUI releases should *not* be used.
Instead, when everything is ready for release, change AC_INIT package
version on `configure.ac`, commit and push all the changes. Then, tag
the latest commit as usual (eg: `git tag v1.1`) and push tags to the
remote repo (`git push --tags`).

The package version on configure.ac and the version tag number (after
the `v`) MUST match, otherwise the generated tarball will not be found.

The tarball will be generated automatically on tag push, and will be
available on the "releases" tab on project page.

Closes: #7