Fix OOM bug when using `diffwr` option
Before this patch, free() was called only when blocks were not the same,
causing memory leaks when they were, what could eventually cause an OOM
due to the accumulated effect of those leaks on large files.

This patch frees allocated memory for both paths, thus fixing the bug.
Thanks @szolnokit for the report.

Closes #16.