configure.ac: remove AC_PROG_GCC_TRADITIONAL
The macro AC_PROG_GCC_TRADITIONAL is obsolete; it was needed for old GNU
C compilers when dealing with buggy ioctl implementations that required
a workaround for their header files.

It has not been needed for more than thirty years now, and was declared
obsolescent as of autoconf 2.60 (ca. 2006). AC_PROG_GCC_TRADITIONAL no
longer does anything and can be removed.